# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.ResourceType:
    properties:
      createdAt:
        type: string
      description:
        type: string
      frameworkId:
        description: Identifier of the framework (e.g., "mc-iam-manager", "mc-infra-manager")
        type: string
      id:
        description: Unique identifier within the framework (e.g., "workspace", "vm")
        type: string
      name:
        description: Display name (e.g., "Workspace", "Virtual Machine")
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: MC IAM Manager API Documentation
  title: MC IAM Manager API
  version: 0.6.2
paths:
...
  /api/resource-types/cloud-resources:
    post:
      consumes:
      - application/json
      description: 새로운 리소스 타입을 생성합니다
      operationId: createResourceType
      parameters:
      - description: Resource Type Info
        in: body
        name: resourceType
        required: true
        schema:
          $ref: '#/definitions/model.ResourceType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ResourceType'
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cloud에서 관리되는 Resource(vm, nlb, k8s 등의 그룹) 새 리소스 타입 생성
      tags:
      - resource-types
  /api/resource-types/cloud-resources/framework/{frameworkId}/id/{resourceTypeId}:
    delete:
      consumes:
      - application/json
      description: 리소스 타입을 삭제합니다
      operationId: deleteResourceType
      parameters:
      - description: Resource Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Resource Type not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 리소스 타입 삭제
      tags:
      - resource-types
    get:
      consumes:
      - application/json
      description: 특정 리소스 타입을 ID로 조회합니다
      operationId: getCloudResourceTypeByID
      parameters:
      - description: Resource Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResourceType'
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Resource Type not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 리소스 타입 ID로 조회
      tags:
      - resource-types
    put:
      consumes:
      - application/json
      description: 리소스 타입 정보를 업데이트합니다
      operationId: updateResourceType
      parameters:
      - description: Resource Type ID
        in: path
        name: id
        required: true
        type: string
      - description: Resource Type Info
        in: body
        name: resourceType
        required: true
        schema:
          $ref: '#/definitions/model.ResourceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResourceType'
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Resource Type not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 리소스 타입 업데이트
      tags:
      - resource-types
  /api/resource-types/cloud-resources/list:
    post:
      consumes:
      - application/json
      description: 모든 리소스 타입 목록을 조회합니다
      operationId: listCloudResourceTypes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ResourceType'
            type: array
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 리소스 타입 목록 조회
      tags:
      - resource-types
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"